wolfram@the-dreams.de:
When checking the commandline-help for generating the manpage from it,
I wondered if CONFIG_NONE is really needed. Removing it, makes IMHO
this upcoming task easier and also simplifies the code, as it has often
to be treated as an exception. I think the information presented there
can also be presented otherwise (like preceeding 'conf-var:').



git-svn-id: http://svn.unix-ag.uni-kl.de/vpnc@226 315857ad-0bdb-0310-b42e-dec37551a5f0