Gerald Hanusch (sonicwall integration ('hacking')) Wolfgang Astleitner (code integration, testing) both: University of Linz / ZID
Disconnecting didn't work fine (compare to VPNC/Known Bugs:
disconnecting does not work reliable with all supported targets
(a work-around is to connect with incorrect password, and then again
with correct password)).
We had a similar problem with Sonicwall: Routing to internal nets did
only work every 2nd time (or when using the 'incorrect
password'-workaround) -> when debugging the communication we learned
that sonicwall didn't like the chained disconnect request but did well
with the request in two separated steps.

We've tested the patched version with a Sonicwall and a Cisco and both
worked fine with it. Maybe this will also solve the problem with other
targets.



git-svn-id: http://svn.unix-ag.uni-kl.de/vpnc@239 315857ad-0bdb-0310-b42e-dec37551a5f0