Wolfram Sang <wolfram@the-dreams.de>
I think a small HISTORY file could be useful (like "when again was
this feature introduced?"). As such information is already on the
webpage, I attached a proposal (simply webpage dumped with elinks and
removed the unneccessary stuff).

Me:

Named the file ChangeLog as this seems to be the convention with
many GNU projects (although we aren't one :)


git-svn-id: http://svn.unix-ag.uni-kl.de/vpnc@244 315857ad-0bdb-0310-b42e-dec37551a5f0