Johan Fischer <jfischer@cmss-systems.com>
We recently upgraded our netscreen ScreenOS firewall from 5.4 to 6.0. vpnc trunk
M284 was working perfectly fine with the previous screenOS 5.4.0, but after the
upgrade, we had intermittent problem during the init of the connection.

The problem seems to be that the vpn endpoint is sending a UDP NAT-keepalive
packet in the middle of the Xauth and breaks the client expecting the xauth
result (which closes the vpn connection).
[...]
This is a similar problem that the SonicWall problem reported in Nov 2007
(http://lists.unix-ag.uni-kl.de/pipermail/vpnc-devel/2007-November/001981.html).

After investigating a bit more the function and flow, it seems I have
misconfiguration of the active NAT setting.

the function recv_ignore_dup is called but because my active_nat is different
than NATT_ACTIVE_DRAFT_OLD, the function is not discarding the packet.

[patch deleted]

The connection is now always successful and the logs are indicating:
   Received UDP Nat-Keepalive bug nat active mode incorrect: 3
   Received UDP Nat-Keepalive bug nat active mode incorrect: 3

3 seems to match NATT_ACTIVE_RFC        /* draft 2 and RFC3947 / RFC3948 */
in tunip.h

So somewhere, either vpnc or the screenOS system negociated a wrong NAT :/

Will check further.

Find attached the full patch to vpnc.c which also includes some new VID and
DEBUG2 messages for the IKE negociation. (some part should be discarded and were
use to debug).



git-svn-id: http://svn.unix-ag.uni-kl.de/vpnc@289 315857ad-0bdb-0310-b42e-dec37551a5f0