Johan Fischer <jfischer@cmss-systems.com>
Following the previous patch about some NAT-T keepalive UDP packets received
during the session setup, it seems the netscreen is also sending those packets
when the tunnel is Idle. With the current code, the vpnc is logging in syslog
some alerts messages about packets too short and dropped.

The attached patch will detect the NAT packet and just log an INFO message in
the syslog (you can remove that logging) and reduce the noise of vpnc during
Idle times. the logic is not changed (still returning -1 when seen) and this is
mostly a cosmetic change.



git-svn-id: http://svn.unix-ag.uni-kl.de/vpnc@290 315857ad-0bdb-0310-b42e-dec37551a5f0