Looking at the seemlingly dpd problems that have shown up over time there seems to be one category that has eluded fixing/working around.
Recent debugs seem to indicate that these are caused by the concen-
trator sending a delete-ipsec-spi payload instead of initiating a
ipsec-rekeying-proces itself. If we continue to use the spi that
the concentrator wanted to delete the concentrator terminates the
connection.

Attempted fix: When we receive a delete-ipsec message we just rerun
the quickmode exchange.

This patch has not been tested by anyone who experienced/reported
the problem so far - it just looks like a possible fix.



git-svn-id: http://svn.unix-ag.uni-kl.de/vpnc@331 315857ad-0bdb-0310-b42e-dec37551a5f0