Nicholas Reilly <njreilly@gmail.com>
... they have the Firewall 
option enabled on the concentrator. I turned on logs whilst doing a 
connection and saw the concentrator sent me the hex 8001 0001 8002 0001 
8003 0002 for the firewall option before disconnecting me. I guessed 
that it looked like a list of firewall options that it wanted so I 
patched the client to send that to it rather than the empty option that 
is currently sent. The concentrator sent back a bit more but I remained 
connected ...



git-svn-id: http://svn.unix-ag.uni-kl.de/vpnc@358 315857ad-0bdb-0310-b42e-dec37551a5f0